Fixed support for DW_AT_ranges attribute:
When addr2line is searching for a function given an address(PC), it
should also consider that the DW_AT_ranges attribute maybe used in
place of DW_AT_lopc/DW_AT_hipc pair to indicate that the function's
address is inside a continuous or non-continuous address ranges. For
example, an inlined function that's inclined multiple times in
different source code locations.